Fix error in distance calculation for equal coordinates.
Distance calculation is now cut short for equal nodes. For some coordinates,
the calculation could result in NaN.

Other cosmetic changes.



git-svn-id: http://svn.mkgmap.org.uk/mkgmap/trunk@877 25d90789-57f7-4ee0-8453-03a3dfeeeb22

+++ src/uk/me/parabola/mkgmap/reader/polish/PolishMapDataSource.java
@@ -487,7 +487,7 @@ public class PolishMapDataSource extends MapperBasedMapDataSource implements Loa
 		if (fields[0].length() == 0)
 			i = 1;
 
-		Float f1 = Float.valueOf(fields[i]);
+		Double f1 = Double.valueOf(fields[i]);
 		Double f2 = Double.valueOf(fields[i+1]);
 		return new Coord(f1, f2);
 	}
